+++ sdio_ext.cpp
@@ -1,6 +1,6 @@
 #include "sdio_ext.h"
 
-namespace cm3ext {
+namespace cm3cpp {
 
 namespace sdio {
