+++ timer_ext.cpp
@@ -25,7 +25,7 @@ TIM C++ Wrapper of libopencm3 library for STM32F2, STM32F4
 
 #include "timer_ext.h"
 
-namespace cm3ext {
+namespace cm3cpp {
 
 namespace tim {
